+++ src/components/token/token-input.component.ts
@@ -1,11 +1,14 @@
 import {
   Component,
-  Input,
-  forwardRef
+  Input, Output,
+  forwardRef,
+  EventEmitter,
+  ContentChildren,
+  QueryList
 } from '@angular/core';
 import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
 
-
+import { TokenComponent } from './token.component';
 
 export const CUSTOM_INPUT_CONTROL_VALUE_ACCESSOR2: any = {
   provide: NG_VALUE_ACCESSOR,
@@ -25,9 +28,22 @@ export const CUSTOM_INPUT_CONTROL_VALUE_ACCESSOR2: any = {
 export class TokenInputComponent implements ControlValueAccessor {
 
   @Input('tokens') tokens: any[] = [];
+  @Output('change') onChange = new EventEmitter();
+  value: any[];
+
+  @ContentChildren(TokenComponent)
+  templateItems: QueryList<TokenComponent>;
 
   addtext: string = '';
 
+
+  ngAfterContentInit() {
+    // transform template-items if available
+    let templateItemsAr = this.templateItems.toArray();
+    if (templateItemsAr.length > 0) {
+      this.tokens = templateItemsAr.map(i => i.toObject());
+    }
+  }
 
   keydown(ev) {
     if (ev.key != 'Enter') return;
@@ -35,11 +51,15 @@ export class TokenInputComponent implements ControlValueAccessor {
 
     this.tokens.push({ label: this.addtext });
     this.addtext = '';
+
+    this.onChange.emit(this.tokens);
+
     !!this.onChangeCallback && this.onChangeCallback(this.tokens);
   }
 
   remove(token) {
     this.tokens = this.tokens.filter(t => t.label != token.label);
+    this.onChange.emit(this.tokens);
     !!this.onChangeCallback && this.onChangeCallback(this.tokens);
   }